feat: Partial reveal/match logic to cards
When you click a card, if it's not already open, then it opens. If you click two cards that match, they will remain open and their background colors will change in order to indicate that on the page. I still don't know why you can't have two unmatched cards open on the page, will have to fix that.