fix: Two unmatching cards together on page
I guess you could say this is also a type of feat - if I click two unmatching cards in succession, they now remain on the page for a little while, so the user can see what they both were. In the past, if you had opened a card, the next card you would click would not open if it was not a match. Also added the mechanism for the victory pop-up, placeholder that changes the heading color for now, will add it later